Remove legacy API key route module
